+++ CHANGELOG.md
@@ -14,6 +14,7 @@ You can view the most detailed change log on the https://github.com/embar-/pdsa-
 ### New features
 - Additional custom Viz engine for plotting graphs (as an alternative to Cytoscape).
 - Ability for the user to specify columns instead of requiring them to be named with standard names.
+- Option to select a references sheet (if there are more than 1).
 
 ## v1.3 (2025-02-10)
 ### Fixes